Refactor customer modules into package
