+++ customers/__init__.py
@@ -1,0 +1,2 @@
+"""Customer-related components for the print shop simulation."""
+
